move db module back to be more backwards compatible
